Refactor audit log command listener and improve logging
Replaces LogCommandListener with AuditLogCommandListener for audit log channel management, updating all references and registration in FireRun. Refactors DatabaseConnector with clearer method names and adds channel/message registration checks. Enhances message logging in MessageLogListener, including improved handling of message edits and deletions, and clarifies audit log mapping in various listeners.
