+++ src/main/java/org/papertrail/cleanup/BotKickListener.java
@@ -26,6 +26,7 @@ public class BotKickListener extends ListenerAdapter {
 		Guild leftGuild = event.getGuild();
 		try {
 			dc.unregisterGuildAndChannel(leftGuild.getId(), TableNames.AUDIT_LOG_TABLE);
+			dc.unregisterGuildAndChannel(leftGuild.getId(), TableNames.MESSAGE_LOG_REGISTRATION_TABLE);
 		} catch (SQLException e) {
 			Logger.error("Could not auto-unregister guild upon guild leave", e);
 		}
