+++ src/main/java/org/papertrail/listeners/guildlisteners/ServerBoostListener.java
@@ -26,6 +26,7 @@ public class ServerBoostListener extends ListenerAdapter {
 	public void onGuildMemberUpdateBoostTime(GuildMemberUpdateBoostTimeEvent event) {
 		
 		// this will return a non-null text id if a channel was previously registered in the database
+		// server boost logs are mapped to audit log table
 		String registeredChannelId=dc.retrieveRegisteredChannelId(event.getGuild().getId(), TableNames.AUDIT_LOG_TABLE);
 
 		if(registeredChannelId==null ||registeredChannelId.isBlank()) {
