+++ src/main/java/org/papertrail/listeners/voicelisteners/GuildVoiceListener.java
@@ -26,6 +26,7 @@ public class GuildVoiceListener extends ListenerAdapter {
 		
 		// this will return a non-null text id if a channel was previously registered in
 		// the database
+		// guild voice events are mapped to audit log table
 		String registeredChannelId = dc.retrieveRegisteredChannelId(event.getGuild().getId(), TableNames.AUDIT_LOG_TABLE);
 
 		if (registeredChannelId == null || registeredChannelId.isBlank()) {
